requirements.txt
dotenv
google
google-genai
python-dotenv
Django
djangorestframework


